fixed bugs for adding new limit in dolla mode and also tag parsing for adding records in dolla mode
