+++ src/main/java/dolla/parser/Parser.java
@@ -3,6 +3,7 @@ package dolla.parser;
 import dolla.ModeStringList;
 import dolla.Time;
 import dolla.ui.EntryUi;
+import dolla.ui.LimitUi;
 import dolla.ui.Ui;
 import dolla.ui.ModifyUi;
 import dolla.command.Command;
@@ -80,7 +81,7 @@ public abstract class Parser implements ParserStringList, ModeStringList {
      * @param str String (of number) to be converted into integer type.
      * @return Integer type of the specified string.
      */
-    public double stringToDouble(String str) {
+    public static double stringToDouble(String str) {
         double newDouble = 0.0;
         try {
             newDouble = Double.parseDouble(str);
@@ -332,4 +333,45 @@ public abstract class Parser implements ParserStringList, ModeStringList {
         return tempStr;
     }
 
+    //@@author Weng-Kexin
+    double findLimitAmount() {
+        double amount = 0;
+        try {
+            amount = stringToDouble(inputArray[2]);
+        } catch (NumberFormatException e) {
+            LimitUi.invalidAmountPrinter();
+        }
+        return amount;
+    }
+
+    private Boolean verifyLimitType(String limitType) {
+        return limitType.equals(LIMIT_TYPE_S)
+                || limitType.equals(LIMIT_TYPE_B);
+    }
+
+    private Boolean verifyLimitDuration(String limitDuration) {
+        return limitDuration.equals(LIMIT_DURATION_D)
+                || limitDuration.equals(LIMIT_DURATION_W)
+                || limitDuration.equals(LIMIT_DURATION_M);
+    }
+
+    private Boolean verifyLimitAmount(double limitAmount) {
+        return (limitAmount != 0);
+    }
+
+    protected Boolean verifySetLimitCommand() {
+        boolean isValid;
+        try {
+            String typeStr = inputArray[1];
+            double amountInt = findLimitAmount();
+            String durationStr = inputArray[3];
+            isValid = verifyLimitType(typeStr) && verifyLimitAmount(amountInt) && verifyLimitDuration(durationStr);
+        } catch (IndexOutOfBoundsException e) {
+            LimitUi.invalidSetCommandPrinter();
+            isValid = false;
+        } catch (Exception e) {
+            isValid = false;
+        }
+        return isValid;
+    }
 }
